Project Start!
The is the beginning of project Save Point! Please read through Github
on how to use it correctly, and be sure to only commit changes that are
CERTAIN to work, and only changes in the Assets or Project Settings
folders. Library and Temp folders are not to be uploaded. Thanks!