Zach - README & Movement
Made it so the character can do all that is listed, and that the game
knows when the player's character is doing each. This will make it
easier when we implement the stealth mechanic to enemies. Depending on
what the player's character is doing is if they notice. I can change
these movement numbers around to make him faster/slower and all.